Dev server 1

+++ server/patients/src/main/java/com/fitbee/patients/repositories/PatientRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 public interface PatientRepository extends JpaRepository<Patient,Integer> {
     Patient findByFirstName(String firstName);
     //  Patient findByName(String name);
-    Patient findByUserUserId(int id);
-    Boolean existsByUserUserId(int id);
+    Patient findByUserUserId(long id);
+    Boolean existsByUserUserId(long id);
 }
